update to original (#1) (#1)
* Add shadow-plugin-tor's continuous integration test

* Add a .clang-format configuration

This specifies the formatting preferences for the clang-format tool,
which should pick up this configuration automatically.  I started with
clang-format's defaults and made a few tweaks in accordance with the
preferences observed in the existing code.

I don't plan or recommend mass reformatting, but this is useful
for keeping modified code tidy and formatted consistently.
A mass reformat might make sense down the line once we've had time
to get comfortable with the options, and don't have any large
pending changes (since it'd make merging a huge headache).

* Experimental implementation of network interface token buckets

* Add router element to buffer data received by interface

The idea is that the router element can support various different
active queue management (AQM) algorithms. I believe such an
algorithm is necessary to smooth out the packet dropping behavior
at the interface in order for TCP to work correctly. CoDel is the
first planned AQM.

* Redesign router and router queue management

The queue manager logic is now separated from the router, so that
we can support multiple queue management algorithms.

Two simple algorithms are supported now: a single packet queue,
and a queue that can grow to a static size.

Also, a queue manager based on CoDel is stubbed out but not yet
functional.

* Avoid assertion error

* Single event to refill token buckets

* Refactor receive packets function

* Implement CoDel and make it the default router queue manager

* Hosts init routers, add router and netiface object counters

* Update expected leakcheck test output

* Additional code comments and cleanup

* Better version management

* Bump version: 1.13.2 → 1.14.0

* Use apt-get instead of apt

apt is meant to be friendlier for human usage but doesn't guarantee a
stable interface. apt-get is meant to be the stable interface used for
scripting.

* Update packages before trying to install.

CI is failing because it's trying to fetch package paths that no longer
exist. We need to update our cache before trying to install packages.

Co-authored-by: Jim Newsome <jnewsome@torproject.org>
Co-authored-by: Rob Jansen <jansen@cs.umn.edu>